nokialcd: added a simple bit_buffer framework
This allows to push bits into a buffer which can then be sent to the lcd
in one transfer.

Note: the maximum amount of bytes in one transfer is limited by the esp32
hardware and is 4094 bytes, this is not enough to send a complete screen
update in one transfer. (132 * 132 * 1.5 * 9 / 8 = 29403 bytes.