nokialcd: use the bit_buffer for images & clearing
A full screen image is sent to the lcd one row at a time. This results
in 132 spi transfers for a full screen update vs 3 transfers per 2
pixels. The screen update is almost instantaneous now.

A similar process is used for clearing the screen but with even more
optimization: one row gets prepared in the bit_buffer, then the same
buffer gets sent 132 times.